Remove redundant model
