+++ lib/ThemeController.php
@@ -37,7 +37,6 @@ class ThemeController extends \TMS\Theme\Base\ThemeController {
 
         add_action( 'init', function () {
             \ArchiveArtist::hooks();
-            \ArchiveArtwork::hooks();
         } );
     }
 }
